refactor(GameSystem): changed class names + method names
GameManager -> RacingGame / GameSetting -> GameSystem validate~ methods names changed
